+++ sdks/python/test-suites/direct/build.gradle
@@ -44,7 +44,8 @@ task ioCrossLanguagePostCommit {
 }
 
 task messagingCrossLanguagePostCommit {
-  getVersionsAsList('cross_language_validates_py_versions').each {
+  // Messaging E2E tests has testcontainers overhead. Single Python version suffices and reducing CI flakiness
+  getVersionsAsList('cross_language_validates_py_versions').take(1).each {
     dependsOn.add(":sdks:python:test-suites:direct:py${getVersionSuffix(it)}:messagingCrossLanguagePythonUsingJava")
   }
 }
